+++ CHANGELOG.md
@@ -15,6 +15,8 @@ Generally, more significant breaking changes will be put near the top of each ca
   - Casts the type of the wrapped value: `maybe(...).cast(int)` → `maybe[int]`
 - Added method `Maybe.flatten`
   - Replicates Rust's [`Option::flatten`](https://doc.rust-lang.org/std/option/enum.Option.html#method.flatten)
+- Added method `Maybe.inspect`
+  - Replicates Rust's [`Option::inspect`](https://doc.rust-lang.org/std/option/enum.Option.html#method.inspect)
 - Added method `Maybe.reduce`
   - Replicates Rust's [`Option::reduce`](https://doc.rust-lang.org/std/option/enum.Option.html#method.reduce)
 - Added method `Maybe.replace`
@@ -23,6 +25,7 @@ Generally, more significant breaking changes will be put near the top of each ca
   - Replicates Rust's [`Option::zip`](https://doc.rust-lang.org/std/option/enum.Option.html#method.zip)
 - Added test `test_maybe.test_cast`
 - Added test `test_maybe.test_flatten`
+- Added test `test_maybe.test_inspect`
 - Added test `test_maybe.test_reduce`
 - Added test `test_maybe.test_replace`
 - Added test `test_maybe.test_zip`